refactoring code style
