+++ src/components/Slider/Slider.jsx
@@ -32,7 +32,7 @@ const Slider = () => {
         mousewheel={true}
         keyboard={true}
         loop={true}
-        autoplay={{ delay: 10000, disableOnInteraction: false }}
+        autoplay={{ delay: 5000, disableOnInteraction: false }}
         modules={[Navigation, Pagination, Mousewheel, Keyboard, Autoplay]}
         className="mySwiper">
         <SwiperSlide>
